Small refactor and type safety suggestions from coderabbit
